+++ nic/apollo/api/impl/artemis/meter_impl_state.hpp
@@ -65,7 +65,7 @@ public:
     /// \brief    return lpm table's size based on IP address family
     /// \param[in] af    IP address family
     /// \return    size of single IPv4/IPv6 meter lpm table's size
-    uint32_t table_size(uint8_t af) const {
+    uint64_t table_size(uint8_t af) const {
         if (af == IP_AF_IPV4) {
             return v4_lpm_table_size_;
         }
@@ -103,10 +103,10 @@ private:
     indexer       *v4_lpm_idxr_;           ///< indexer to allocate v4 lpm mem block
     indexer       *v6_lpm_idxr_;           ///< indexer to allocate v6 lpm mem block
     mem_addr_t    v4_lpm_region_addr_;     ///< base address for the v4 lpm region
-    uint32_t      v4_lpm_table_size_;      ///< size of each v4 lpm table
+    uint64_t      v4_lpm_table_size_;      ///< size of each v4 lpm table
     uint32_t      v4_lpm_max_prefixes_;    ///< max IPv4 routes per meter
     mem_addr_t    v6_lpm_region_addr_;     ///< base address for the v6 lpm region
-    uint32_t      v6_lpm_table_size_;      ///< size of each v6 lpm table
+    uint64_t      v6_lpm_table_size_;      ///< size of each v6 lpm table
     uint32_t      v6_lpm_max_prefixes_;    ///< max IPv6 routes per meter
     indexer       *policer_idxr_;          ///< indexer to allocate policer indices
     indexer       *stats_idxr_;            ///< indexer to allocate accounting stats indices
