+++ nic/sdk/platform/utils/mpartition.hpp
@@ -38,8 +38,8 @@ namespace utils {
  */
 typedef struct mpartition_region_s {
     char            mem_reg_name[MEM_REG_NAME_MAX_LEN];   /**< Name */
-    uint32_t        size;           /**< Size */
-    uint32_t        block_size;     /**< Size of each block, if exists */
+    uint64_t        size;           /**< Size */
+    uint64_t        block_size;     /**< Size of each block, if exists */
     uint32_t        max_elements;   /**< max. no. of elements, if exists */
     mem_addr_t      start_offset;   /**< Start address offset */
     cache_pipe_t    cache_pipe;     /**< Cached pipe */
@@ -116,7 +116,7 @@ public:
      *
      * @return #Memory mpartition region size in bytes
      */
-    uint32_t size(const char *name);
+    uint64_t size(const char *name);
 
     /**
      * @brief Get size of each block in mpartition region
@@ -125,7 +125,7 @@ public:
      *
      * @return #Memory size of each block in mpartition region
      */
-    uint32_t block_size(const char *name);
+    uint64_t block_size(const char *name);
 
     /**
      * @brief Get number of elements in mpartition region
